基本功能

+++ AQL/src/main/java/www/aql/com/entity/response/BaseResponse.java
@@ -4,7 +4,7 @@ package www.aql.com.entity.response;
  * Created by Jason on 2016/7/17.
  */
 public class BaseResponse {
-    public static final int SUCCESS_OK = 0;
+    public static final long SUCCESS_OK = 0;
     public long errcode;
     public String errmsg;
